Rename gql operation
